Create a working mechanism for users to create additional views.
Introduce BranchListView, an extension of ListView that is only
available to MultiBranchProjects.  The view has an additional URL
binding for "branch" that is necessary for job links to work.  Standard
views do not have this binding, so links to branch jobs from standard
non-primary views end up broken.  This view is marked as
non-instantiable so it does not show outside of MultiBranchProjects.

Build the message from MultiBranchProjectWelcomeView into
BranchListView by overriding the inherited noJob.jelly.  This allows us
to remove the primary view override and also more accurately displays
configuration prompts to the user.

Add newView.jelly to MultiBranchProject, overriding the inherited view.
The new view now uses MultiBranchProject#getViewDescriptors() to
determine views that users are allowed to create.

Add a migration path for views that deletes all non-compatible view
types and adds a BranchListView if all views were deleted.

Override view initialization to use BranchListView instead of AllView.