[FIXED JENKINS-37219] Add a job property for overriding branch index triggers.
Works in both directions - disables branch index triggers if set to
false, enables them if set to true and either there's a
NoTriggerBranchProperty applying to the branch or there's a
NoTriggerOrganizationFolderProperty applying to the organization.

Theoretically, the config shouldn't show up for any job type but
multibranch jobs, where it'll only show up in the Snippet Generator anyway.