externalize common NoTriggerProperty logic to an abstract class
The `NoTriggerBranchProperty` and `NoTriggerOrganizationFolderProperty` classes have to provide the same features. The logic was copy-pasted between the classes, so the code was duplicated and hard to maintain (unwritten contract). The logic has been externalized to a new class `NoTriggerMultiBranchQueueDecisionHandler`. It must be extended by the `Dispatcher` class. Methods required by the `Dispatcher` class are defined in the `NoTriggerProperty` class which has to be extended by the property classes.

The `NoTriggerOrganizationFolderProperty` class provides the `branches` property. It determines for which branches the automatic builds should not be suppressed. It has been added to the `NoTriggerBranchProperty` class. To make the logic backward compatible, it by default matches no branches (`^$`).

To implement the parametrized tests easily, the `JUnitParams` library has been added. It has API very similar to JUnit 5, so it will be very easy to migrate to it in the future.