[JENKINS-63673] Allow configuring suppressed events in NoTrigger properties (#244)
* [JENKINS-63673] Allow configuring suppressed events in NoTriggerBranchProperty

The branch property was by default suppressing all triggers: indexing and events (e.g. webhooks). The indexing mechanism is problematic for cloud environments when Jenkins configuration is not kept on a persistent volume. Every Jenkins creation schedules builds of all branches and pull requests which may block the instance for many hours or even days. This change introduces a configuration option which allows administrators/developers to decide if both or only one trigger's kind should be suppressed. It is useful for the mentioned cloud environments because the indexing mechanism could be disabled without ignoring webhooks.

* add upgrade (migration) tests

* Update src/main/resources/jenkins/branch/Messages.properties

Co-authored-by: Jesse Glick <jglick@cloudbees.com>

* add shouldSuppress method to make code more readable

* remove @since as requested

* externalize common NoTriggerProperty logic to an abstract class

The `NoTriggerBranchProperty` and `NoTriggerOrganizationFolderProperty` classes have to provide the same features. The logic was copy-pasted between the classes, so the code was duplicated and hard to maintain (unwritten contract). The logic has been externalized to a new class `NoTriggerMultiBranchQueueDecisionHandler`. It must be extended by the `Dispatcher` class. Methods required by the `Dispatcher` class are defined in the `NoTriggerProperty` class which has to be extended by the property classes.

The `NoTriggerOrganizationFolderProperty` class provides the `branches` property. It determines for which branches the automatic builds should not be suppressed. It has been added to the `NoTriggerBranchProperty` class. To make the logic backward compatible, it by default matches no branches (`^$`).

To implement the parametrized tests easily, the `JUnitParams` library has been added. It has API very similar to JUnit 5, so it will be very easy to migrate to it in the future.

* correct set variable

* get OverrideIndexTriggersJobProperty from job

The `OverrideIndexTriggersJobProperty` class is a job property, so it must be get from the `job` object instead of the branch which is represented by the job.

* correct grammar in help files

* add static modifier when possible

The `static` modifier informs developers that the method doesn't change the state of the object, so it could improve readability.

* correct validation method name

Co-authored-by: Jesse Glick <jglick@cloudbees.com>