[JENKINS-74879] Use strikethrough on merged pull request name column
https://issues.jenkins.io/browse/JENKINS-74879 reports that the
strikethrough text is no longer previously displayed when a pull request
has been merged or closed.  The strikethrough text is a nice visual
indicator that the job is no longer active.  Let's bring it back so that
the visual hint is still there for users.

The strikethrough text is displayed on the job page, but not in the
table of jobs for each pull request or branch.

Change was originally introduced when pull request titles were enabled
by default in the column name.  Refer to:

* https://github.com/jenkinsci/branch-api-plugin/pull/476

Assumed that it is better to drop the jenkins-table__link formatting in
the plugin rather than adjust the CSS definion in Jenkins core.  That CSS
definition has been in place in Jenkins core since Sep 2022 and likely
would have unexpected side effects if changed in core.  Refer to:

* https://github.com/jenkinsci/jenkins/pull/6248

Testing done:

* Confirmed that the link has strikethrough text with this change.
* Confirmed that the non-striketrhrough link still looks good