add a BaseView the ensures DefaulViewsTabBar is used
With https://github.com/jenkinsci/jenkins/pull/10379 it is possible for
users to define a default views tabbar. When a user chooses a different
implementation e.g. from favorite-views plugin, this affects also the
views in org and multibranch jobs as they use the ListView from core.

This change introduces a BaseView that brings it's own main.jelly
ensuring the viewstabbar from the corresponding viewholder is used.
The BaseView also unifies the almost identical view implementations.